+++ src/LiteGraph.Sdk/VectorSearchDomainEnum.cs
@@ -1,0 +1,21 @@
+﻿namespace LiteGraph.Sdk
+{
+    /// <summary>
+    /// Vector search domain.
+    /// </summary>
+    public enum VectorSearchDomainEnum
+    {
+        /// <summary>
+        /// Graph.
+        /// </summary>
+        Graph,
+        /// <summary>
+        /// Node.
+        /// </summary>
+        Node,
+        /// <summary>
+        /// Edge.
+        /// </summary>
+        Edge
+    }
+}