+++ src/LiteGraph/Graph.cs
@@ -51,6 +51,11 @@
         /// </summary>
         public object Data { get; set; } = null;
 
+        /// <summary>
+        /// Vectors.
+        /// </summary>
+        public List<VectorMetadata> Vectors { get; set; } = null;
+
         #endregion
 
         #region Private-Members